Bugfix: ensure `tool` is set to conda for conda-installed packages (#5624)
# Description

When setting the `tool` (pip or conda), only the dropdown widget for
to-be-installed packages was being checked, so it was only used for
**installation** of packages.
This PR ensures that we check also check the *source* of already
installed packages when determining the `tool` to be used.
This ensures that condo-installed packages are uninstalled using
conda—the same goes for upgrading.

## Type of change
- [x] Bug-fix (non-breaking change which fixes an issue)

# References
closes https://github.com/napari/napari/issues/5623

# How has this been tested?
Locally
- Uninstall after conda install
- Upgrading after conda install:

## Final checklist:
- [x] My PR is the minimum possible work for the desired functionality