Bugfix (macOS): Remove PYTHONEXECUTABLE from `env` for conda installs (#5622)
# Description
In https://github.com/napari/napari/pull/5531 PYTHONEXECUTABLE is set to
`sys.executable` on macOS to fix an issue with venv/pyenv and launching
napari.
It seems to have a side effect that makes conda plugin installations
using the new Plugin Manager not work when napari is launched from the
Terminal on macOS, probably because `sys.executable`, is in the env that
napari was run from. But that env doesn't have conda/mamba, only `base`
does.

So this change removes PYTHONEXECUTABLE from the QProcessEnvironment
within the `CondaInstallerTool` class.

## Type of change
<!-- Please delete options that are not relevant. -->
- [x] Bug-fix (non-breaking change which fixes an issue)

# References
Similar issue over at spyder:
https://github.com/spyder-ide/spyder/issues/20599
PR that added the PYTHONEXECUTABLE: #5531 cc: @aganders3
Closes #5621 5621

# How has this been tested?
tested locally on my mac

## Final checklist:
- [x] My PR is the minimum possible work for the desired functionality
- [x] I have commented my code, particularly in hard-to-understand areas