Add warning notification when connectivity errors occur in when opening the plugin manager (#6931)
Fixes https://github.com/napari/napari-plugin-manager/issues/26

# Description

Catches connectivity errors when fetching for plugin information when
opening the plugin manager.

Although the issue lives in the plugin manager repo, the code that
fetches the code still lives in napari, hence with the PR seem to be
better resolved on this side.

# Result

## Before this PR

![napari-before](https://github.com/napari/napari/assets/3627835/6e704463-2dd3-4701-9aa5-93f2501e48fd)

## With this PR

![napari](https://github.com/napari/napari/assets/3627835/ffa3ad0a-da18-4e76-a4aa-8a007f90ce36)

The warning reads:

>There seems to be an issue with network connectivity. Remote plugins
cannot be installed, only local ones.