Use npe2 adaptor when discovering plugins for the installed list (#131)
Prior to this PR, installing an npe1 plugin with Use npe2 adaptor turned on in napari didn't recognize it as a shimmed plugin. As a result, it didn't ask the user to restart (even though a restart is needed to load the plugin), and didn't display the npe2 (adapted) badge.

This PR updates the manager to read the state of the setting and use it appropriately.

To test that this PR works:

Install napari from main, launch a viewer
Turn on Use npe2 adaptor in Preferences. Restart napari (skip this if you already had the setting turned on)
Open the plugin manager using Install/Uninstall plugins
Install an npe1 plugin e.g. Label-Creator or anything off this list
❌ Plugin installs, you're not told to restart, the plugin doesn't get the (npe1) adapted label, the plugin widget is not listed in Plugins
(Optional) Restart napari, plugin gets the (npe1) adapted label, plugin widget is listed in Plugins
Uninstall your plugin
Install napari-plugin-manager from this branch
Launch a viewer, install your plugin again
✔️ You're told to restart, plugin gets (npe1) adapted label, and upon restarting, your widget is listed
---------

Co-authored-by: Draga Doncila <ddon0001@student.monash.edu>
Co-authored-by: pre-commit-ci[bot] <66853113+pre-commit-ci[bot]@users.noreply.github.com>
Co-authored-by: Juan Nunez-Iglesias <juan.nunez-iglesias@monash.edu>