Don't rely on `use_npe2_adaptor` setting (#191)
Partner PR to [napari/#8448](https://github.com/napari/napari/pull/8448).

This PR updates the usage of the `use_npe2_adaptor` setting so that if there **is** a value for the setting, we use it, otherwise we set it to `True`. This ensures old versions of napari with the latest version of `napari-plugin-manager` respect the setting, and new versions of `napari` behave correctly.

I have updated the tests to always use the npe2 adaptor, but I have not removed bits of code or tests that may now be obsolete. We'll have to do a pass at this eventually, but I'd prefer to wait for @jaimergp and @dalthviz to be around for this, as we may want to update our mock plugin managers with more plugins.

---------

Co-authored-by: Draga Doncila <ddon0001@student.monash.edu>
Co-authored-by: Grzegorz Bokota <bokota+github@gmail.com>